Rename folders to lowercase and move images to res
Components and other folders were using capital letter and/or camel
case formats, renamed them to lowercase. Also move images to a "res"
(resources) folder.
