+++ src/components/homepage/index.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import './Homepage.css';
-import Post from '../Post'
-import TextsRepository from '../../TextsRepository'
+import Post from '../post'
+import TextsRepository from '../../texts_repository'
 
 let textsRepository = new TextsRepository();
 
@@ -26,7 +26,7 @@ class Homepage extends Component {
   render() {
     let data;
     if (this.state.loading) {
-      data = <img  src={ require('../../images/loading.gif') } />
+      data = <img  src={ require('../../res/images/loading.gif') } />
     } else {
         let { posts } = this.state;
         data = posts.map((post, i) => <Post marked_text={post} key={i} />)
