+++ src/texts_repository/index.js
